interim commit, cleanup before code review

+++ transitclock/src/main/java/org/transitclock/avl/ApcDataProcessor.java
@@ -241,7 +241,7 @@ public class ApcDataProcessor {
     apcAggregator.analyze(matches);
   }
 
-  public synchronized int cacheSize() {
+  public synchronized long cacheSize() {
      return apcAggregator.cacheSize();
   }
